Note/app/requirements.txt
flask
google-cloud-firestore
flask-wtf
flask-restful
wtforms